+++ src/datasources/data-frame/constants.ts
@@ -24,3 +24,5 @@ export const errorCodes: { [key: number]: string } = {
 };
 
 export const propertiesCacheTTL = 1000 * 60 * 5; // 5 minutes
+
+export const TAKE_LIMIT = 1000;